refactor: rename term print module
