+++ pulsar/src/pulsar/term_print.rs
